EEGToolkit.jl

Computational EEG analysis with emphasis in sleep neuroscience.

Package Features

Loading EEG data

EEG visualization

Sleep stage handling and NREM period detection

Power spectral analysis

Spindle detection algorithms

Function Documentation

EEGToolkit.epoch — Function

Returns a vector [i₁, …, iₖ] with all indexes corresponding to the nth epoch of the EEG. The default sampling rate is used to compute the indexes.

Returns a vector [i₁, …, iₖ] with all indexes corresponding to epochs n, n+1, …, m of the EEG. The default sampling rate is used to compute the indexes.

Returns a vector [x₁, …, xₖ] with all values of the signal channel in the nth epoch.

Returns a vector [x₁, …, xₖ] with all values of the signal channel in the epochs n, n+1, …, m.

EEGToolkit.get_stage_indexes — Function

This function maps an EEG and astages vector to the array of all indexes whose values in an EEG signal pertain to a stage in stages.

EEGToolkit.plot_eeg_overlay — Function

Plots with the active backend all EEG channels in channels in the range from the nth to the mth epoch.

EEGToolkit.Spectrogram — Type

Structure for spectrogram estimation. Estimations are by default one-sided, with frequencies ranging from [0, fₛ/2]. The signal is split into possibly overlapping windows of length L; within each window, Welch's method is used to compute the PSD with overlapping windows. For Barlett's method, one can set the inner window length and the overlap to zero.

Fields

time::Vector : Time domain (x)

freq::Vector{<:AbstractFloat}: Frequency domain (y)

spectrums::Matrix{<:AbstractFloat}: Power spectrum (z). Rows are time and columns are frequency.

segment_length::Integer : Length of each segment in time.

Constructors

Spectrogram(signal::Vector{<:AbstractFloat}, fs::Integer, segmentlength::Integer, overlap::AbstractFloat, innerwindowlength::Integer, inneroverlap::AbstractFloat) : Computes the Spectrogram of a signal with sampling rate fs in windows of length segment_length (in number of samples) with a certain overlap ∈ [0, 1].

Within each window, the PSD constructor is used to compute either a Welch or a Barlett method estimation, depending on the inner_window_length and inner_overlap parameters. An optional normalization factor and a zero-padding length can be included, as in the PSD constructor.

EEGToolkit.gen_time_domain — Function

Given an EEG, generates the time vector t₁, …, tₙ corresponding to EEG signals from time s to e.

EEGToolkit.next_power_of_two — Function

Given an integer n, finds the least m = 2ᵏ s.t. m ≥ n.

EEGToolkit.get_stage — Function

Returns the subset of an EEG channel corresponding to a stage.

EEGToolkit.overlaps — Function

Splits a vector v into segments of length L with an overlap overlap_frac expressed as a fraction of L.

EEGToolkit.AmplitudeSpectrum — Type

Structure for amplitude spectrum estimations. Estimations are by default one sided, with frequencies ranging from [0, fₛ/2].

Fields

freq::Vector{<:AbstractFloat}: Frequency range of the spectrum

spectrum::Vector{<:AbstractFloat}: Estimated spectral amplitude

formula::String: A string representation of the formula used for the estimation.

Constructors

AmplitudeSpectrum(x::Vector{<:AbstractFloat}, samplingrate::Integer, pad::Integer) : Computes a direct PSD over a signal x with a given `samplingrate`.

EEGToolkit.PSD — Type

Structure for PSD estimations. Estimations are by default one sided, with frequencies ranging from [0, fₛ/2].

Fields

freq::Vector{<:AbstractFloat}: Frequency range of the spectrum

spectrum::Vector{<:AbstractFloat} : Estimated spectral density in dB.

method::String: Estimation method used

formula::String : A string representation of the formula used for the estimation.

Constructors

PSD(x::Vector, samplingrate::Integer, pad::Integer = 0): Computes a direct PSD over a signal x with a given `samplingrate`.

PSD(x::Vector, fs::Int, L::Int, overlap::Union{ <:AbstractFloat, Integer }, normalization::Union{ <:AbstractFloat, Integer } = 1): Splits the signal x into segments of length L with an overlap in [0, 1). The overlap is understood to be a fraction of the segment length. PSD is estimated within and averaged across all segments. If overlap is zero, this results in Barlett's method. If overlap is greater than zero, this results in Welch's method. If pad is zero no zero-padding is done. If pad is greater than zero, each segment is zero-padded to a length of pad.

EEGToolkit.plot_spectrogram — Function

Plots a spectogram spec either in 2d (type = 1) or 3d (type = 2). An optional frequency limit (freq_lim) may be set (defaults to 30Hz). The color palette color may be set; defaults to nipy_spectral.

EEGToolkit.zero_pad — Function

Zero-pads a numeric vector v to a desired_length

EEGToolkit.freq_band — Function

Given a PSD, returns a Vector{AbstractFloat} with the powers within the frequency band [lower, upper].

Given a spectrogram, returns a Vector{<:AbstractFloat} with the powers within a frequency band [lower, upper] of a specific window (row of the spectrogram).

Given a spectrogram, returns a Matrix{<:AbstractFloat} with the powers within a frequency band [lower, upper] across all windows.

EEGToolkit.plot_eeg — Function

Plots with the active backend all EEG channels in channels in the range from the nth to the mth epoch.

Plots with the active backend all EEG channels in channels in the range from the nth to the mth epoch.

EEGToolkit.artifact_reject — Function

Given an EEG, a 2x2 matrix associating epoch-subepoch pairs with artifacts, and a signal, returns a subset of the signal with all sub-epochs containing artifacts removed.

The signal is split in epoch-length windows and each window is split in subepoch-length windows; the matrix gives the epoch and subepoch indexes to be removed.