EEGToolkit.jl

Computational EEG analysis with emphasis in sleep neuroscience.

Package Features

Loading EEG data

EEG visualization

Sleep stage handling and NREM period detection

Power spectral analysis

Spindle detection algorithms

Function Documentation

EEGToolkit.EEG — Type

A mutable struct representing EEG data with associated metadata.

Fields

signals::Dict{String, Vector{<:AbstractFloat}}: A dictionary mapping signal labels (strings) to arrays of floating-point values.

sampling_rates::Dict{String, Integer}: A dictionary mapping signals (strings) to the integer sampling rates.

fs::Integer: A default sampling rate that will be used for calculations. Defaults to the maximum sampling rate among all signals.

epoch_length::Integer: Length of each epoch (in seconds).

staging::Vector{String}: A vector of stage labels corresponding to each epoch.

id::String: An identifier for the EEG.

Constructors

EEG(file, fs, epoch_length, staging): Constructs an EEG object from an EDF file (file) containing EEG data. The function reads the signals, computes the necessary metadata (fs, N, epoch_count), and initializes the EEG struct with the provided staging vector.

Example

staging_vector = CSV.read("path/to/stage_data/eeg_staging.csv") # A vector with a stage per each epoch in the EEG
eeg_data = EEG("path/to/edf_data/data.edf", 30, staging_vector)

# Alternatively, if no stage data exists, it is safe to do 
eeg_data = EEG("path/to/edf_data/data.edf", 30, [])

# or simply 
eeg_data = EEG("path/to/edf_data/data.edf", 30) # Staging vector defaults to [] if not provided

# or simply
eeg_data = EEG("path/to/edf_data/data.edf") # Epoch length defaults to 30 if not provided.

EEGToolkit.epoch — Function

epoch(eeg::EEG, n::Integer, fs::Integer=-1)

Returns a vector [i₁, …, iₖ] with all indexes corresponding to the nth epoch of the EEG. The default sampling rate is used to compute the indexes.

epoch(eeg::EEG, n::Integer, m::Integer, fs::Integer=-1)

Returns a vector [i₁, …, iₖ] with all indexes corresponding to epochs n, n+1, …, m of the EEG. The default sampling rate is used to compute the indexes.

epoch(eeg::EEG, n::Integer, channel::String)

Returns a vector [x₁, …, xₖ] with all values of the signal channel in the nth epoch.

epoch(eeg::EEG, n::Integer, m::Integer, channel::String)

Returns a vector [x₁, …, xₖ] with all values of the signal channel in the epochs n, n+1, …, m.

EEGToolkit.get_stage_indexes — Function

get_stage_indexes(eeg::EEG, stages::Vector)

This function maps an EEG and astages vector to the array of all indexes whose values in an EEG signal pertain to a stage in stages.

EEGToolkit.plot_eeg_overlay — Function

plot_eeg_overlay(eeg::EEG, channels::Vector{String}, n::Integer, m::Integer)

Plots with the active backend all EEG channels in channels in the range from the nth to the mth epoch.

EEGToolkit.Spectrogram — Type

Structure for spectrogram estimation. Estimations are by default one-sided, with frequencies ranging from [0, fₛ/2]. The signal is split into possibly overlapping windows of length L; within each window, Welch's method is used to compute the PSD with overlapping windows. For Barlett's method, one can set the inner window length and the overlap to zero.

Fields

time::Vector : Time domain (x)

freq::Vector{<:AbstractFloat}: Frequency domain (y)

spectrums::Matrix{<:AbstractFloat}: Power spectrum (z). Rows are time and columns are frequency.

segment_length::Integer : Length of each segment in time.

Constructors

Spectrogram(signal::Vector{<:AbstractFloat}, fs::Integer, segment_length::Integer, overlap::AbstractFloat, inner_window_length::Integer, inner_overlap::AbstractFloat) : Computes the Spectrogram of a signal with sampling rate fs in windows of length segment_length (in number of samples) with a certain overlap ∈ [0, 1].

Within each window, the PSD constructor is used to compute either a Welch or a Barlett method estimation, depending on the inner_window_length and inner_overlap parameters. An optional normalization factor and a zero-padding length can be included, as in the PSD constructor.

EEGToolkit.gen_time_domain — Function

gen_time_domain(eeg::EEG, s::Union{AbstractFloat,Integer}, e::Union{AbstractFloat,Integer}, fs::Integer=-1)

Given an EEG, generates the time vector t₁, …, tₙ corresponding to EEG signals from time s to e.

EEGToolkit.next_power_of_two — Function

next_power_of_two(n::Int)

Given an integer n, finds the least m = 2ᵏ s.t. m ≥ n.

EEGToolkit.get_stage — Function

function get_stage(eeg::EEG, channel::String, stages::Vector)

Returns all portions of an EEG channel in a given stage of the staging vector.

EEGToolkit.overlaps — Function

overlaps(v::Vector{T}, L::Int, overlap_frac::Union{Float64,Int}) where {T}

Splits a vector v into segments of length L with an overlap overlap_frac expressed as a fraction of L.

EEGToolkit.AmplitudeSpectrum — Type

Structure for amplitude spectrum estimations. Estimations are by default one sided, with frequencies ranging from [0, fₛ/2].

The formula used is

\[\left[\frac{2|H(f)|}{\sum_i w_i} \]

with $w_i$ a Hanning window.

Fields

freq::Vector{<:AbstractFloat}: Frequency range of the spectrum

spectrum::Vector{<:AbstractFloat}: Estimated spectral amplitude

formula::String: A string representation of the formula used for the estimation.

Constructors

AmplitudeSpectrum(x::Vector{<:AbstractFloat}, sampling_rate::Integer, pad::Integer) : Computes a direct PSD over a signal x with a given sampling_rate.

EEGToolkit.PSD — Type

Structure for PSD estimations. Estimations are by default one sided, with frequencies ranging from [0, fₛ/2].

The default formula is

\[\frac{2|H(f)|^2}{f_s \sum_i w_i^2}\]

with $w_i$ a Hanning window. This means the estimation is normalized by the sampling rate by default. This can be changed by setting the normalization parameter equal to $\frac{1}{f_s}$, canceling out the factor in the denominator.

If Barlett or Welch's mehtod is used (i.e. if the second constructor is used), the formula becomes formula = "1/(M * normalization) ∑ ᵢᴹ [ 2|Hᵢ(f)|² / ( fₛ ∑ wᵢ² ) ] where w₁, …, wₗ a Hanning window, M the number of segments, and Hᵢ(f) the FFT of the ith segment of the signal. "

\[\frac{1}{M \varphi} \sum_i^M \left[ \frac{2|H_i(f)|^2}{f_s \sum_i w_i^2} \right]\]

where $\varphi$ is an optional normalization factor defined by the normalization parameter (defaults to 1).

Fields

freq::Vector{<:AbstractFloat}: Frequency range of the spectrum

spectrum::Vector{<:AbstractFloat} : Estimated spectral density in dB.

method::String: Estimation method used

formula::String : A string representation of the formula used for the estimation.

Constructors

PSD(x::Vector, sampling_rate::Integer, pad::Integer = 0): Computes a direct PSD over a signal x with a given sampling_rate.

PSD(x::Vector, fs::Int, L::Int, overlap::Union{ <:AbstractFloat, Integer }, normalization::Union{ <:AbstractFloat, Integer } = 1): Splits the signal x into segments of length L with an overlap in [0, 1). The overlap is understood to be a fraction of the segment length. PSD is estimated within and averaged across all segments. If overlap is zero, this results in Barlett's method. If overlap is greater than zero, this results in Welch's method. If pad is zero no zero-padding is done. If pad is greater than zero, each segment is zero-padded to a length of pad.

EEGToolkit.plot_spectrogram — Function

plot_spectrogram(spec::Spectrogram, freq_lim::AbstractFloat=30.0, type::Int=1, color=:nipy_spectral)

Plots a spectogram spec either in 2d (type = 1) or 3d (type = 2). An optional frequency limit (freq_lim) may be set (defaults to 30Hz). The color palette color may be set; defaults to nipy_spectral.

EEGToolkit.zero_pad — Function

zero_pad(v::Vector{<:AbstractFloat}, desired_length::Integer) where {T}

Zero-pads a numeric vector v to a desired_length

EEGToolkit.freq_band — Function

freq_band(spec::Union{PSD,AmplitudeSpectrum}, lower::AbstractFloat, upper::AbstractFloat)

Given a PSD, returns a Vector{AbstractFloat} with the powers within the frequency band [lower, upper].

freq_band(spec::Spectrogram, lower::AbstractFloat, upper::AbstractFloat, window::Integer)

Given a spectrogram, returns a Vector{<:AbstractFloat} with the powers within a frequency band [lower, upper] of a specific window (row of the spectrogram).

freq_band(spec::Spectrogram, lower::AbstractFloat, upper::AbstractFloat)

Given a spectrogram, returns a Matrix{<:AbstractFloat} with the powers within a frequency band [lower, upper] across all windows.

EEGToolkit.plot_eeg — Function

Plots with the active backend all EEG channels in channels in the range from the nth to the mth epoch.

plot_eeg(eeg::EEG, channels::Vector{String}, n::Integer, m::Integer)

Plots with the active backend all EEG channels in channels in the range from the nth to the mth epoch.

EEGToolkit.artifact_reject — Function

artifact_reject(eeg::EEG, anom_matrix::Matrix, signal::String)

Given an EEG, a 2x2 matrix associating epoch-subepoch pairs with artifacts, and a signal, returns a subset of the signal with all sub-epochs containing artifacts removed.

The signal is split in epoch-length windows and each window is split in subepoch-length windows; the matrix gives the epoch and subepoch indexes to be removed.