EEGToolkit.jl

Computational EEG analysis with emphasis in sleep neuroscience.

Developed at the Laboratory for the Study of Sleep Slow-wave activity

The Gods of the earth and sea Sought thro' Nature to find this Tree, But their search was all in vain: There grows one in the Human Brain.

William Blake

This package has three aims in mind:

Simplicity

Transparency

Efficiency

Simplicity means that a scientist with little programming background should be able to use it. Transparency means that any methodology implemented by the package should be transparent enough so as to be reported in a scientific paper. Efficiency means that large EEGs (e.g. sleep EEGs) should be processed and analyzed in minutes or less.

Transparency affects primarily spectral analysis. Most packages (e.g. the gsignal package in R) do not report how power spectrum analysis is done. Power spectral analysis is not standardized and may be computed in several many ways. This difficults comparing and testing results. This package reports the formula used for PSD estimation and incorporates features to extract such details from any PSD object.

Lastly, this

Package Features

Loading and processing EEG data

EEG visualization

Sleep stage handling

NREM Period detection

Power spectral analysis

Spindle detection algorithms

EEG

EEGToolkit.EEG — Type

A mutable struct representing EEG data with associated metadata.

Fields

signals::Dict{String, Vector{<:AbstractFloat}}: A dictionary mapping signal labels (strings) to arrays of floating-point values.

sampling_rates::Dict{String, Integer}: A dictionary mapping signals (strings) to the integer sampling rates.

fs::Integer: A default sampling rate that will be used for calculations. Defaults to the maximum sampling rate among all signals.

epoch_length::Integer: Length of each epoch (in seconds).

staging::Vector{String}: A vector of stage labels corresponding to each epoch.

id::String: An identifier for the EEG.

Constructors

EEG(file, fs, epoch_length, staging): Constructs an EEG object from an EDF file (file) containing EEG data. The function reads the signals, computes the necessary metadata (fs, N, epoch_count), and initializes the EEG struct with the provided staging vector.

Example

staging_vector = CSV.read("path/to/stage_data/eeg_staging.csv") # A vector with a stage per each epoch in the EEG
eeg_data = EEG("path/to/edf_data/data.edf", 30, staging_vector)

# Alternatively, if no stage data exists, it is safe to do 
eeg_data = EEG("path/to/edf_data/data.edf", 30, [])

# or simply 
eeg_data = EEG("path/to/edf_data/data.edf", 30) # Staging vector defaults to [] if not provided

# or simply
eeg_data = EEG("path/to/edf_data/data.edf") # Epoch length defaults to 30 if not provided.

EEGToolkit.epoch — Function

epoch(eeg::EEG, n::Integer, fs::Integer=-1)

Returns a vector [i₁, …, iₖ] with all indexes corresponding to the nth epoch of the EEG. The default sampling rate is used to compute the indexes.

epoch(eeg::EEG, n::Integer, m::Integer, fs::Integer=-1)

Returns a vector [i₁, …, iₖ] with all indexes corresponding to epochs n, n+1, …, m of the EEG. The default sampling rate is used to compute the indexes.

epoch(eeg::EEG, n::Integer, channel::String)

Returns a vector [x₁, …, xₖ] with all values of the signal channel in the nth epoch.

epoch(eeg::EEG, n::Integer, m::Integer, channel::String)

Returns a vector [x₁, …, xₖ] with all values of the signal channel in the epochs n, n+1, …, m.

EEGToolkit.get_stage_indexes — Function

get_stage_indexes(eeg::EEG, stages::Vector)

This function maps an EEG and astages vector to the array of all indexes whose values in an EEG signal pertain to a stage in stages.

EEGToolkit.plot_eeg_overlay — Function

plot_eeg_overlay(eeg::EEG, channels::Vector{String}, n::Integer, m::Integer)

Plots with the active backend all EEG channels in channels in the range from the nth to the mth epoch.

EEGToolkit.gen_time_domain — Function

gen_time_domain(eeg::EEG, s::Union{AbstractFloat,Integer}, e::Union{AbstractFloat,Integer}, fs::Integer=-1)

Given an EEG, generates the time vector t₁, …, tₙ corresponding to EEG signals from time s to e.

EEGToolkit.get_stage — Function

function get_stage(eeg::EEG, channel::String, stages::Vector)

Returns all portions of an EEG channel in a given stage of the staging vector.

EEGToolkit.overlaps — Function

overlaps(v::Vector{T}, L::Int, overlap_frac::Union{Float64,Int}) where {T}

Splits a vector v into segments of length L with an overlap overlap_frac expressed as a fraction of L.

EEGToolkit.plot_spectrogram — Function

plot_spectrogram(spec::Spectrogram, freq_lim::AbstractFloat=30.0, type::Int=1, color=:nipy_spectral)

Plots a spectogram spec either in 2d (type = 1) or 3d (type = 2). An optional frequency limit (freq_lim) may be set (defaults to 30Hz). The color palette color may be set; defaults to nipy_spectral.

EEGToolkit.freq_band — Function

freq_band(spec::Union{PSD,AmplitudeSpectrum}, lower::AbstractFloat, upper::AbstractFloat)

Given a PSD, returns a Vector{AbstractFloat} with the powers within the frequency band [lower, upper].

freq_band(spec::Spectrogram, lower::AbstractFloat, upper::AbstractFloat, window::Integer)

Given a spectrogram, returns a Vector{<:AbstractFloat} with the powers within a frequency band [lower, upper] of a specific window (row of the spectrogram).

freq_band(spec::Spectrogram, lower::AbstractFloat, upper::AbstractFloat)

Given a spectrogram, returns a Matrix{<:AbstractFloat} with the powers within a frequency band [lower, upper] across all windows.

EEGToolkit.plot_eeg — Function

Plots with the active backend all EEG channels in channels in the range from the nth to the mth epoch.

plot_eeg(eeg::EEG, channels::Vector{String}, n::Integer, m::Integer)

Plots with the active backend all EEG channels in channels in the range from the nth to the mth epoch.

EEGToolkit.artifact_reject — Function

artifact_reject(eeg::EEG, anom_matrix::Matrix, signal::String)

Given an EEG, a 2x2 matrix associating epoch-subepoch pairs with artifacts, and a signal, returns a subset of the signal with all sub-epochs containing artifacts removed.

The signal is split in epoch-length windows and each window is split in subepoch-length windows; the matrix gives the epoch and subepoch indexes to be removed.

NREM Period detection

Following Feinberg & Floyed and Dijk, a NREM period is a sequence of epochs satisfying the following two conditions:

It starts with stages 2, 3 or 4.

It contains at least 15 minutes of stages 2, 3 or 4 exist in total (not necessarily continuously).

It ends with either $a.$ at least 5 minutes of REM or $b.$ at least 5 minutes of wakefulness.

The sequence is allowed to contain ocurrences of REM sleep or wakefulness, as long as their duration is inferior to 5 minutes. Naturally, the epochs corresponding to these occurrences will not be signaled as NREM.

Importantly, the restriction that ending REM periods must last at least 5 minutes is not imposed when detecting the first and the last NREM period in a night of sleep.

Let $n$ be the number of epochs corresponding to $15$ minutes and $m$ the number of epochs corresponding to $5$ minutes. (In 30 second epohcs, $n = 30, m = 10$).

For this algorithm to work, the staging field of an EEG object must have been set to a stage vector $\vec{s}$, and this vector must contain only the strings $1, \ldots, 6, ?$ (with $5$ marking REM, $6$ wakefulness, $?$ unknown/unstaged).

The algorithm works by mapping $\vec{s}$ to $\alpha = s_1 \ldots s_q$ a word over the language generated by $\Sigma = \{1, \ldots, 6\}$.

Observe that the language $[(5+6)^*(2+3+4)^*]^*$ is partitioned into $U$ and $U’$, where $U$ is the set of words containing at least $n$ symbols $2, 3, 4$ where neither $5$ nor $6$ occur consecutively $m = 10$ times. Then $\alpha$ can be decomposed into

\[\alpha = \psi_1 \phi_1 \psi_2 \phi_2 \ldots \psi_k \phi_k \psi_{k+1}\]

where $\\phi_i = \\varphi_i (5^m5^* + 6^m6^*)$ and $\\varphi_i \\in U$. Such a decomposition readily provides the number of NREM periods in the EEG (i.e. $k$). Furthermore, the epochs which comprise these periods are easily inferable from the decomposition.

EEGToolkit.nrem — Function

nrem(staging::Vector, n::Integer=30, m::Integer=10)

Finds the k underlying NREM periods in a staging vector. Returns a vector of vectors V s.t. the ith vector in V contains the epochs which comprise the ith NREM period. Thus, the length of V is k the number of NREM periods.

The staging field of the EEG must have been set to a vector containing only the symbols 1, …, 6, ? where 5 denotes REM, 6 denotes wakefulness, and ? denotes unscored/unstaged.

nrem(eeg::EEG, n::Integer=30, m::Integer=10)

Finds the k underlying NREM periods in the staging vector of an EEG.

Spindle detection

This package implements two spindle detection algorithms discussed in O'Reilly and Nielsen (2015). We give a brief overview of them here but refer to their original publications for further detail.

EEGToolkit.sigma_index — Function

function sigma_index(x::Vector{<:AbstractFloat}, fs::Integer)

The $\Sigma$-index algorithm (Huupponen et al., 2007) find spindles by detecting abnormally high amplitudes values among the spindle frequency band. Per each 1 second window of the EEG,

the maximum amplitude in the spindle frequency, which we call $S_{max}$

the average amplitude in the low alpha and theta frequencies, which we call

$\alpha_{mean}, \theta_{mean}$

the maximum alpha amplitude $\alpha_{max}$

are all computed. The $\Sigma$-index is defined to be zero if $\alpha_max > S_{max}$, and otherwise

\[f(S_{max}, \alpha_{mean}, \phi_{mean}) = \frac{2S_{max}}{ \alpha_{mean} + \phi_{mean} } \]

Higher values are indicative of a higher spindle probability. The rejection threshold recommended in the original paper is $\lambda = 4.5$.

EEGToolkit.relative_spindle_power — Function

relative_spindle_power(x::Vector{<:AbstractFloat}, fs::Integer)

The Relative Spindle Power (RSP) algorithm (Devuyst et al., 2011) also detects abnormal values along the spindle frequency band. For every 1 second window, the amplitude spectrum $S(t)$ is computed, and the RSP is defined as

\[RSP(t) = \frac{\int_{11}^{16} S(t, f) df}{\int_{0.5}^{40} S(t, f) df}\]

This definition is more intelligible than the that of the sigma index, insofar as it represents the ratio of the total power in the spindle band with respect to the total power in the $\delta$ to $\phi$ frequency range. It is evident by definition that $0 \leq RSP \leq 1$. Higher values are indicative of a higher spindle probability (it should be clear that $RSP$ is not a probability itself). The rejection threshold recommended in the original paper is $\lambda = 0.22$.

Power spectral analysis

EEGToolkit.AmplitudeSpectrum — Type

Structure for amplitude spectrum estimations. Estimations are by default one sided, with frequencies ranging from [0, fₛ/2].

The formula used is

\[\frac{2|H(f)|}{\sum_i w_i}\]

with $w_i$ a Hanning window.

Fields

freq::Vector{<:AbstractFloat}: Frequency range of the spectrum

spectrum::Vector{<:AbstractFloat}: Estimated spectral amplitude

formula::String: A string representation of the formula used for the estimation.

Constructors

AmplitudeSpectrum(x::Vector{<:AbstractFloat}, sampling_rate::Integer, pad::Integer) : Computes a direct PSD over a signal x with a given sampling_rate.

EEGToolkit.PSD — Type

Structure for PSD estimations. Estimations are by default one sided, with frequencies ranging from [0, fₛ/2].

The default formula is

\[\frac{2|H(f)|^2}{f_s \sum_i w_i^2}\]

with $w_i$ a Hanning window. This means the estimation is normalized by the sampling rate by default. This can be changed by setting the normalization parameter equal to $\frac{1}{f_s}$, canceling out the factor in the denominator.

If Barlett or Welch's mehtod is used (i.e. if the second constructor is used), the formula becomes

\[\frac{1}{M \varphi} \sum_i^M \left[ \frac{2|H_i(f)|^2}{f_s \sum_i w_i^2} \right]\]

where $w_1, \ldots, w_n$ a Hanning window, $M$ the number of segments, $H_i(f)$ the FFT of the $i$th segment of the signal, and $\varphi$ an optional normalization factor defined by the normalization parameter (defaults to 1).

Fields

freq::Vector{<:AbstractFloat}: Frequency range of the spectrum

spectrum::Vector{<:AbstractFloat} : Estimated spectral density in dB.

method::String: Estimation method used

formula::String : A string representation of the formula used for the estimation.

Constructors

PSD(x::Vector, sampling_rate::Integer, pad::Integer = 0): Computes a direct PSD over a signal x with a given sampling_rate.

PSD(x::Vector, fs::Int, L::Int, overlap::Union{ <:AbstractFloat, Integer }, normalization::Union{ <:AbstractFloat, Integer } = 1): Splits the signal x into segments of length L with an overlap in [0, 1). The overlap is understood to be a fraction of the segment length. PSD is estimated within and averaged across all segments. If overlap is zero, this results in Barlett's method. If overlap is greater than zero, this results in Welch's method. If pad is zero no zero-padding is done. If pad is greater than zero, each segment is zero-padded to a length of pad.

EEGToolkit.Spectrogram — Type

Structure for spectrogram estimation. Estimations are by default one-sided, with frequencies ranging from [0, fₛ/2]. The signal is split into possibly overlapping windows of length L; within each window, Welch's method is used to compute the PSD with overlapping windows. For Barlett's method, one can set the inner window length and the overlap to zero.

Fields

time::Vector : Time domain (x)

freq::Vector{<:AbstractFloat}: Frequency domain (y)

spectrums::Matrix{<:AbstractFloat}: Power spectrum (z). Rows are time and columns are frequency.

segment_length::Integer : Length of each segment in time.

Constructors

Spectrogram(signal::Vector{<:AbstractFloat}, fs::Integer, segment_length::Integer, overlap::AbstractFloat, inner_window_length::Integer, inner_overlap::AbstractFloat) : Computes the Spectrogram of a signal with sampling rate fs in windows of length segment_length (in number of samples) with a certain overlap ∈ [0, 1].

Within each window, the PSD constructor is used to compute either a Welch or a Barlett method estimation, depending on the inner_window_length and inner_overlap parameters. An optional normalization factor and a zero-padding length can be included, as in the PSD constructor.

Spectrogram(segs::Vector{Vector{<:AbstractFloat}}, fs::Integer, segment_length::Integer, overlap::AbstractFloat, normalization::Union{AbstractFloat,Integer}=1, pad::Integer=0): This constructor does not take a signal but a split or windowed signal. This is useful, for example, when the EEG is split into windows corresponding to a specific period (e.g. NREM epochs). In this case, each time-instance in the Spectrogram corresponds to one of the windows. Aside from this, there is no difference with the previous constructor.

Helpers

EEGToolkit.next_power_of_two — Function

next_power_of_two(n::Int)

Given an integer n, finds the least m = 2ᵏ s.t. m ≥ n.

EEGToolkit.zero_pad — Function

zero_pad(v::Vector{<:AbstractFloat}, desired_length::Integer) where {T}

Zero-pads a numeric vector v to a desired_length

Example

This is an example script for computing the mean $\delta$ (delta) power in each of the NREM periods of a sleep EEG. We will use the C3 channel.

# First, import the package
using EEGToolkit 
using CSV

# Assuming we have the stage data in a .csv
staging = CSV.read("my_staging_data.csv")

# We read an EEG that has channels C3-A2 and F3-A1
eeg = EEG(edf_file, staging.STAGES)

# Detect the NREM periods
nrems = nrem(eeg)

# Filter the EEG 
eeg.signals["C3-A2"] = filt(digitalfilter(Lowpass(70, fs=eeg.fs), Butterworth(4)), eeg.signals["C3-A2"] )
eeg.signals["C3-A2"] = filt(digitalfilter(Highpass(0.3, fs=eeg.fs), Butterworth(4)), eeg.signals["C3-A2"] )
eeg.signals["C3-A2"] = filt(digitalfilter(Bandstop(50, 70, fs=eeg.fs), Butterworth(4)), eeg.signals["C3-A2"] )

# Split the C3 signal into 30-second windows (not-overlapping).
epochs = overlaps(eeg.signals["C3-A2"], eeg.fs * 30 * 30, 0)

mean_delta_powers = []
for nrem_period in nrems
    # Extract the portion of the signal corresponding to this NREM period
    nrem_epohcs = epochs[nrem_period]
    # Compute its spectrogram. (Observation: This is an example of when it's useful to be 
    # able to pass a vector argument to the Spectrogram constructor.)
    spec = Spectrogram(nrem_epochs, eeg.fs, eeg.fs*5 - 1, 0.5)

    # Extract delta band (0.5 to 3.9 Hz) from the spectrogram.
    delta_band = freq_band(spec, 0.5, 3.9)

    # Compute the mean power in this delta band
    mean_delta_power = mean(delta_band, dims=2) 
    push!(mean_delta_powers, mean_delta_power)
end

# Now the ith element in `mean_delta_powers` is the mean delta power 
of the ith NREM period.